requirements.txt
google-cloud-storage
streamlit
